refactor: se agruparon componentes en sus carpetas correspondientes
