--- src/components/UI/atoms/Input.css
@@ -1,12 +1,0 @@
-input{
-    border-radius: 10px;
-    width: 250px;
-    padding-left: 10px;
-    height: 25px;
-}
-.name{
-    margin: 10px;   
-}
-.name input{
-    width: 400px;
-}
