+++ src/components/UI/atoms/Navbar/Navbar.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../../atoms/Button/Button';
+import Button from '../Button/Button';
 //import { Link } from 'react-router-dom';
 import "./Navbar.css";
