+++ src/components/UI/molecules/CobroEfectivo/BotonCobrarEfectivo.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../atoms/Button';
+import Button from '../../atoms/BotonAc/BotonAc';
 
 
 function PagarEfectivoButton({ onClick }) {
