+++ src/components/UI/molecules/VehicleRegistration.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import Input from '../atoms/Input';
-import Button from '../atoms/Button';
+import Input from '../atoms/Input/Input';
+import Button from '../atoms/BotonAc/BotonAc';
 import './VehicleRegistration.css';
 function VehicleRegistration({ onRegister }) {
     const [matricula, setMatricula] = useState('');
