+++ src/components/UI/organisms/CobroEfectivo/CobroEfectivo.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import PagarEfectivoButton from '../molecules/PagarEfectivoButton';
-import PagoEfectivoForm from '../molecules/PagoEfectivoForm';
+import PagarEfectivoButton from '../../molecules/CobroEfectivo/BotonCobrarEfectivo';
+import PagoEfectivoForm from '../../molecules/CobroEfectivo/FormCobroEfectivo';
 
 function CobroEfectivoSection({ onPagoEfectivo }) {
   const [mostrarFormulario, setMostrarFormulario] = useState(false);
